Included e-mail in User create
-Cleaned up index.js
-Resolved security issue again
