+++ models/user.js
@@ -11,6 +11,12 @@ const UserSchema = new mongoose.Schema({
   password: {
     type: String,
     required: true
+  },
+  email: {
+    type: String,
+    unique: false,
+    required: false,
+    trim: true
   }
 })
