+++ views/register.ejs
@@ -11,6 +11,7 @@
         <form action="/register" method="post" onsubmit="submitBtn.disabled = true">
           <input name="username" type="text" class="ghost-input" placeholder="Username" minlength="3" required>
           <input name="password" type="text" class="ghost-input" placeholder="Password" minlength="3" required>
+          <input name="email" type="text" class="ghost-input" placeholder="Email" minlength="3" required>
           <input type="submit" class="ghost-button" value="Create account" name = "submitBtn">
         </form>
       </fieldset>
